+++ bstrlen.c
@@ -3,6 +3,6 @@
 #include "uint32.h"
 #include "str.h"
 
-int bstrlen(const char* a) {
+size_t bstrlen(const char* a) {
   if (*a) return str_len(a); else return uint32_read(a+1);
 }
