+++ fmt_asn1int.c
@@ -1,7 +1,7 @@
 #include <asn1.h>
 
-unsigned int fmt_asn1int(char* dest,enum asn1_tagclass tc,enum asn1_tagtype tt,enum asn1_tag tag,unsigned long l) {
-  unsigned int len,tmp;
+size_t fmt_asn1int(char* dest,enum asn1_tagclass tc,enum asn1_tagtype tt,enum asn1_tag tag,unsigned long l) {
+  size_t len,tmp;
   /* first the tag */
   if (!dest) return fmt_asn1tag(0,tc,tt,tag)+1+fmt_asn1intpayload(0,l);
   len=fmt_asn1tag(dest,tc,tt,tag);
