+++ fmt_ldapava.c
@@ -1,8 +1,8 @@
 #include "asn1.h"
 #include "ldap.h"
 
-unsigned int fmt_ldapava(char* dest,struct AttributeValueAssertion* a) {
-  long sum,l;
+size_t fmt_ldapava(char* dest,struct AttributeValueAssertion* a) {
+  size_t sum,l;
   sum=fmt_ldapstring(dest,&a->desc);
   if (dest) dest+=sum;
   l=fmt_ldapstring(dest,&a->value);
