+++ fmt_ldapmessage.c
@@ -1,8 +1,8 @@
 #include "asn1.h"
 #include "ldap.h"
 
-unsigned int fmt_ldapmessage(char* dest,long messageid,long op,long len) {
-  unsigned int l,l2,l3;
+size_t fmt_ldapmessage(char* dest,long messageid,long op,size_t len) {
+  size_t l,l2,l3;
   l2=fmt_asn1INTEGER(0,messageid);
   l3=fmt_asn1transparent(0,APPLICATION,CONSTRUCTED,op,len);
   l=fmt_asn1SEQUENCE(dest,len+l2+l3);
