+++ fmt_ldapstring.c
@@ -1,6 +1,6 @@
 #include "asn1.h"
 #include "ldap.h"
 
-unsigned int fmt_ldapstring(char* dest,struct string* s) {
+size_t fmt_ldapstring(char* dest,struct string* s) {
   return fmt_asn1OCTETSTRING(dest,s->s,s->l);
 }
