+++ mduptab.h
@@ -14,5 +14,5 @@ typedef struct mduptable {
 
 void mduptab_init(mduptab_t* t);
 void mduptab_init_reuse(mduptab_t* t,mstorage_t* s);
-long mduptab_add(mduptab_t* t,const char* s,unsigned int len);
+long mduptab_add(mduptab_t* t,const char* s,size_t len);
 long mduptab_adds(mduptab_t* t,const char* s);
