+++ scan_ldapmessage.c
@@ -1,9 +1,9 @@
 #include "asn1.h"
 #include "ldap.h"
 
-unsigned int scan_ldapmessage(const char* src,const char* max,
-		     unsigned long* messageid,unsigned long* op,unsigned long* len) {
-  unsigned int res,tmp;
+size_t scan_ldapmessage(const char* src,const char* max,
+			unsigned long* messageid,unsigned long* op,size_t* len) {
+  size_t res,tmp;
   if (!(res=scan_asn1SEQUENCE(src,max,len))) goto error;
   if (!(tmp=scan_asn1INTEGER(src+res,max,(long*)messageid))) goto error;
   res+=tmp;
