+++ scan_ldapresult.c
@@ -1,10 +1,10 @@
 #include "asn1.h"
 #include "ldap.h"
 
-unsigned int scan_ldapresult(const char* src,const char* max,unsigned long* result,
-		    struct string* matcheddn,struct string* errormessage,
-		    struct string* referral) {
-  unsigned int res,tmp;
+size_t scan_ldapresult(const char* src,const char* max,unsigned long* result,
+		       struct string* matcheddn,struct string* errormessage,
+		       struct string* referral) {
+  size_t res,tmp;
   if (!(res=scan_asn1ENUMERATED(src,max,result))) return 0;
   if (!(tmp=scan_ldapstring(src+res,max,matcheddn))) return 0;
   res+=tmp;
