+++ scan_ldapstring.c
@@ -1,6 +1,6 @@
 #include <asn1.h>
 #include <ldap.h>
 
-unsigned int scan_ldapstring(const char* src,const char* max,struct string* s) {
+size_t scan_ldapstring(const char* src,const char* max,struct string* s) {
   return scan_asn1STRING(src,max,&s->s,&s->l);
 }
