+++ strstorage.h
@@ -2,5 +2,5 @@
  * later.  On the plus side, the allocation overhead is close to zero.
  * Will return a pointer to the stored copy of the string. */
 
-const char* strstorage_add(const char* s,int n);
+const char* strstorage_add(const char* s,size_t n);
